requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
pypdf>=4.0.0
